+++ editor_patch/event.h
@@ -56,7 +56,8 @@ enum class AlpineDedEventID : int
     AF_Heal,
     Anchor_Marker_Orient,
     Light_State,
-    World_HUD_Sprite
+    World_HUD_Sprite,
+    Set_Light_Color
 };
 
 constexpr int af_ded_event_to_int(AlpineDedEventID event_id) noexcept