+++ game_patch/object/event_alpine.cpp
@@ -80,6 +80,7 @@ FunHook<int(const rf::String* name)> event_lookup_type_hook{
                 {"Anchor_Marker_Orient", 133},
                 {"Light_State", 134},
                 {"World_HUD_Sprite", 135},
+                {"Set_Light_Color", 136},
             };
 
             auto it = custom_event_ids.find(name->c_str());
@@ -141,6 +142,7 @@ FunHook<rf::Event*(int event_type)> event_allocate_hook{
                 {133, []() { return new rf::EventAnchorMarkerOrient(); }},
                 {134, []() { return new rf::EventLightState(); }},
                 {135, []() { return new rf::EventWorldHUDSprite(); }},
+                {136, []() { return new rf::EventSetLightColor(); }},
             };
 
             // find type and allocate
@@ -206,6 +208,7 @@ FunHook<void(rf::Event*)> event_deallocate_hook{
                 {133, [](rf::Event* e) { delete static_cast<rf::EventAnchorMarkerOrient*>(e); }},
                 {134, [](rf::Event* e) { delete static_cast<rf::EventLightState*>(e); }},
                 {135, [](rf::Event* e) { delete static_cast<rf::EventWorldHUDSprite*>(e); }},
+                {136, [](rf::Event* e) { delete static_cast<rf::EventSetLightColor*>(e); }},
             };
 
             // find type and deallocate
@@ -245,7 +248,8 @@ bool is_forward_exempt(rf::EventType event_type) {
         rf::EventType::When_Picked_Up,
         rf::EventType::Set_Entity_Flag,
         rf::EventType::Light_State,
-        rf::EventType::World_HUD_Sprite
+        rf::EventType::World_HUD_Sprite,
+        rf::EventType::Set_Light_Color
     };
 
     return forward_exempt_ids.find(event_type) != forward_exempt_ids.end();
@@ -614,6 +618,18 @@ static std::unordered_map<rf::EventType, EventFactory> event_factories {
                 event->scale = params.float1;
                 event->sprite_filename = params.str1;
                 event->sprite_filename_blue = params.str2;
+            }
+            return event;
+        }
+    },
+    // Set_Light_Color
+    {
+        rf::EventType::Set_Light_Color, [](const rf::EventCreateParams& params) {
+            auto* base_event = rf::event_create(params.pos, rf::event_type_to_int(rf::EventType::Set_Light_Color));
+            auto* event = dynamic_cast<rf::EventSetLightColor*>(base_event);
+            if (event) {
+                event->light_color = params.str1;
+                event->randomize = params.bool1;
             }
             return event;
         }